Various unrelated fixes to kimai update script (#4549)
* Ensure lsb-release is installed

Fixes 4548

* Change to use same syntax as check for composer

There's a proposed solution to #4247 (which doesn't seem to have been committed - I'll fix that in a moment). I've changed the test I'm using to match that suggestion.

* Re-install composer

See #4550

* Try to catch download errors

During troubleshooting of another error, I ran into a condition where downloading the new version failed. It seems that this is because the earlier call to get the $RELEASE had failed.

This change is an attempt to catch download errors, and alert the user, and to not rm the existing install until we know the download worked.

* Re-install php-fpm

I missed in the earlier commit that this wasn't getting reinstalled